+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* **\[core\]** Generate an invariant failure if `ComputeableValue.reportPossiblyChanged()` is invoked on
+  disposed `ComputeableValue` instance.
+
 ### [v0.119](https://github.com/arez/arez/tree/v0.119) (2018-12-07)
 [Full Changelog](https://github.com/arez/arez/compare/v0.118...v0.119)